Reference

ITensorQuantumOperatorDefinitions.siteinds — Method

siteinds(f::Function, N::Integer; kwargs...)

Create an array of N physical site indices where the site type at site n is given by f(n) (f should return a string).

ITensorQuantumOperatorDefinitions.siteinds — Method

siteinds(d::Integer, N::Integer; kwargs...)

Create an array of N site indices, each of dimension d.

Keywords

addtags::String: additional tags to be added to all indices

ITensorQuantumOperatorDefinitions.siteinds — Method

siteinds(tag::String, N::Integer; kwargs...)

Create an array of N physical site indices of type tag. Keyword arguments can be used to specify quantum number conservation, see the space function corresponding to the site type tag for supported keyword arguments.

Example

N = 10
s = siteinds("S=1/2", N; conserve_qns=true)

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"Boson";
      dim = 2,
      conserve_qns = false,
      conserve_number = false,
      qnname_number = "Number")

Create the Hilbert space for a site of type "Boson".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"Electron";
      conserve_qns = false,
      conserve_sz = conserve_qns,
      conserve_nf = conserve_qns,
      conserve_nfparity = conserve_qns,
      qnname_sz = "Sz",
      qnname_nf = "Nf",
      qnname_nfparity = "NfParity")

Create the Hilbert space for a site of type "Electron".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"Fermion";
      conserve_qns=false,
      conserve_nf=conserve_qns,
      conserve_nfparity=conserve_qns,
      qnname_nf = "Nf",
      qnname_nfparity = "NfParity",
      qnname_sz = "Sz",
      conserve_sz = false)

Create the Hilbert space for a site of type "Fermion".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"Qubit")

Create the Hilbert space for a site of type "Qubit".

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"Qudit")

Create the Hilbert space for a site of type "Qudit".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"tJ";
      conserve_qns = false,
      conserve_sz = conserve_qns,
      conserve_nf = conserve_qns,
      conserve_nfparity = conserve_qns,
      qnname_sz = "Sz",
      qnname_nf = "Nf",
      qnname_nfparity = "NfParity")

Create the Hilbert space for a site of type "tJ".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"S=1";
      conserve_qns = false,
      conserve_sz = conserve_qns,
      qnname_sz = "Sz")

Create the Hilbert space for a site of type "S=1".

Optionally specify the conserved symmetries and their quantum number labels.

ITensorQuantumOperatorDefinitions.space — Method

space(::SiteType"S=1/2";
      conserve_qns = false,
      conserve_sz = conserve_qns,
      conserve_szparity = false,
      qnname_sz = "Sz",
      qnname_szparity = "SzParity")

Create the Hilbert space for a site of type "S=1/2".

Optionally specify the conserved symmetries and their quantum number labels.